Fix mobile scroll bugs in browse-view. ItemList is now split in ItemList and ScrollableItemList. That also fixes the buttons appearing in mobile browse-view. Moreover changed top-level css to exclude top and bottom bar from any scroll-behaviour
