+++ projects/bp-gallery/src/components/common/ScrollableItemList.scss
@@ -1,0 +1,23 @@
+@import 'src/shared/style';
+
+.item-list {
+  position: relative;
+}
+
+.button-container {
+  display: flex;
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-content: space-between;
+
+  .MuiIconButton-root {
+    background: $primaryColor !important;
+    color: white;
+    box-shadow: 0px 0px 8px 0px rgba(0, 0, 0, 76%);
+    z-index: 99;
+  }
+}
