+++ projects/bp-gallery/src/components/top-and-bottom-bar/NavigationBar.scss
@@ -12,6 +12,9 @@
   z-index: 99;
   min-height: 0;
   overflow-x: auto;
+  position: fixed;
+  bottom: 0;
+  left: 0;
 
   .nav-element {
     padding: 0.5rem 1rem;
