+++ projects/bp-gallery/src/components/top-and-bottom-bar/TopBar.scss
@@ -11,6 +11,9 @@
   align-items: center;
   width: calc(100% - 4rem);
   z-index: 99;
+  position: fixed;
+  top: 0;
+  left: 0;
 
   .bh-logo {
     img {
